Fixed problem with error messages not being called on connection loss
When connection was lost, Anuglar could not load the error message
partial, so the modal was not being shown, and so users could continue
to interact without their work being saved. Replacing the templateUrl
property with the template property containing the HTML for the modal in
the error modal options fixed this.